Apply correct authorization to application routes

+++ libs/gamify/applications/src/lib/applications.controller.ts
@@ -1,4 +1,4 @@
-import { BadRequestException, Body, Controller, Delete, Get, NotFoundException, Param, ParseIntPipe, Post, Put, SetMetadata, UnauthorizedException, UseGuards } from '@nestjs/common';
+import { BadRequestException, Body, Controller, Delete, Get, NotFoundException, Param, ParseIntPipe, Post, Put, UnauthorizedException } from '@nestjs/common';
 import { ApplicationsService } from './applications.service';
 import { CreateApplicationInput } from './dto/create-application.input';
 import { UpdateApplicationInput } from './dto/update-application.input';
@@ -10,39 +10,28 @@ export class ApplicationsController {
   constructor(private applicationsService: ApplicationsService) {}
 
   @Post()
-  async create(@Body() createApplicationInput: CreateApplicationInput) {
-    console.log(JSON.stringify(createApplicationInput));
-
+  @Roles(Role.MODERATOR)
+  async create(@Body() createApplicationInput: CreateApplicationInput, @User() user: UserModel) {
     if (! await this.applicationsService.isNameUnique(createApplicationInput.name)) {
       throw new BadRequestException("Name must be unique");
     }
 
-    // TODO Validate current user id, should be obtained from current authenticated user.
-    createApplicationInput.ownerUserId = 1;
+    createApplicationInput.ownerUserId = user.id;
 
     return await this.applicationsService.create(createApplicationInput);
   }
 
   @Get()
-  findAll(@User() user: UserModel) {
-    console.log('request user:');
-    console.log(user);
+  findAll() {
     return this.applicationsService.findMany();
   }
 
   @Get(':id')
-  @Roles(Role.ADMIN, Role.MODERATOR)
-  async findOne(@Param('id', ParseIntPipe) id: number, @User() user: UserModel) {
-    if (! await this.applicationsService.canModerateApplication(id, user.id)) {
-      throw new UnauthorizedException('Can not moderate app');
-    }
-
+  async findOne(@Param('id', ParseIntPipe) id: number) {
     return this.findApplicationOrFail(id);
   }
 
   @Put(':id')
-  // Admin      - always
-  // Moderator  - if you own the application, or are added as moderator (future feature)
   @Roles(Role.ADMIN, Role.MODERATOR)
   async update(@Param('id', ParseIntPipe) id: number, @Body() updateApplicationInput: UpdateApplicationInput, @User() user: UserModel) {
     if (! await this.applicationsService.canModerateApplication(id, user.id)) {
@@ -59,7 +48,12 @@ export class ApplicationsController {
   }
 
   @Delete(':id')
-  async remove(@Param('id', ParseIntPipe) id: number) {
+  @Roles(Role.ADMIN, Role.MODERATOR)
+  async remove(@Param('id', ParseIntPipe) id: number, @User() user: UserModel) {
+    if (! await this.applicationsService.canModerateApplication(id, user.id)) {
+      throw new UnauthorizedException();
+    }
+
     await this.findApplicationOrFail(id);
 
     return await this.applicationsService.remove(id);
